:truck: refactor: seperate dto to request, response (#13)
