+++ src/domain/family/dto/photo/create-photo.dto.ts
