+++ src/domain/family/dto/photo/update-photo.dto.ts
