+++ src/domain/family/service/family-member.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreateFamilyMemberDto } from '../dto/request/create-family-member.dto';
-import { UpdateFamilyMemberDto } from '../dto/request/update-family-member.dto';
+import { CreateFamilyMemberDto, UpdateFamilyMemberDto } from '../dto';
 
 @Injectable()
 export class FamilyMemberService {
