+++ src/domain/family/service/post.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreatePostDto } from '../dto/request/create-post.dto';
-import { UpdatePostDto } from '../dto/request/update-post.dto';
+import { CreatePostDto, UpdatePostDto } from '../dto';
 
 @Injectable()
 export class PostService {
